Atualização peso/gestor

See merge request Gerdau/metas-system-management!53

+++ Metas.Application/DTO/GIndicadorDTTO.cs
@@ -11,7 +11,7 @@ namespace Metas.Application.DTO
         public string DESCRICAOINDICADOR { get; set; }
         public int IDUNIDADEMEDIDA { get; set; }
         public int IDFREQUENCIA { get; set; }
-        public decimal PESO { get; set; }
+        public int PESO { get; set; }
         public decimal MINIMO { get; set; }
         public decimal PLANEJADO { get; set; }
         public decimal DESAFIO { get; set; }
